requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
python-dotenv==1.0.1
schedule==1.2.1
beautifulsoup4==4.12.3
requests==2.31.0
gql==3.4.1
aiohttp==3.9.3
requests-oauthlib==2.0.0 